backend/requirements.txt
# AstroGuide Agent Dependencies

# Google Agent Development Kit
google-adk>=1.0.0

# Astronomical image analysis (for annotator.py)
astroquery>=0.4.6
astropy>=5.0
numpy>=1.21.0
Pillow>=9.0.0
python-dotenv>=1.0.0

# Webcam capture
opencv-python>=4.5.0

# API Server
fastapi>=0.100.0
uvicorn>=0.22.0

# Google GenAI SDK (for direct Gemini API calls)
google-genai>=1.0.0
